+++ Data/DeriveLiftedInstances/Internal.hs
@@ -30,8 +30,8 @@ import Data.Traversable (for)
 -- @
 -- meth0 :: a
 -- meth1 :: Int -> a
--- meth2 :: a -> Either Bool a -> Int
--- meth3 :: Maybe [a] -> a
+-- meth2 :: a -> Either Bool a -> Sum Int
+-- meth3 :: Maybe [a] -> IO a
 -- @
 --
 -- the resulting template haskell declarations are (pseudo code):
@@ -39,36 +39,34 @@ import Data.Traversable (for)
 -- @
 -- meth0 = $res ($op "meth0" meth0)
 -- meth1 a = $res (($op "meth1" meth1) `$ap` ($arg Int a))
--- meth2 v0 v1 = (($op "meth2" meth2) `$ap` ($var fold0 v0)) `$ap` ($var fold1 v1)
---   where
---     fold0 _ f = f
---     fold1 map f = [| $map $f |]
--- meth3 v2 = $res (($op "meth2" meth2) `$ap` ($var fold2 v2))
---   where
---     fold2 map f = [| $map ($map $f) |]
+-- meth2 ($inp v0) ($inp v1) = $cst (($op "meth2" meth2) `$ap` ($var (`iterate` 0) v0)) `$ap` ($var (`iterate` 1) v1)
+-- meth3 ($inp v2) = $eff (($op "meth2" meth2) `$ap` ($var (`iterate` 2) v2))
 -- @
 data Derivator = Derivator {
   res :: Q Exp -> Q Exp, -- ^ Convert the result of the method
+  cst :: Q Exp -> Q Exp, -- ^ Convert the result of the method if it is a constant
+  eff :: Q Exp -> Q Exp, -- ^ Convert the result of the method if it has an effect
   op  :: Name -> Q Exp -> Q Exp, -- ^ Convert the method (still unapplied to any arguments)
   arg :: Type -> Q Exp -> Q Exp, -- ^ Convert an argument
   var :: (Q Exp -> Q Exp -> Q Exp) -> Q Exp -> Q Exp, -- ^ Convert a variable
+  inp :: Q Pat -> Q Pat, -- ^ Generate an input pattern for a variable
   ap  :: Q Exp -> Q Exp -> Q Exp -- ^ Apply an argument or variable to the method
 }
 
 varExp :: Name -> Q Exp
 varExp = pure . VarE
 
-varPat :: Name -> Q Pat
-varPat = pure . VarP
-
 -- | The identity `Derivator`. Not useful on its own, but often used as input for other `Derivator`s.
 idDeriv :: Derivator
 idDeriv = Derivator {
-  res  = id,
-  op   = const id,
-  arg  = const id,
-  var  = const id,
-  ap   = \f a -> [| $f $a |]
+  res = id,
+  cst = id,
+  eff = id,
+  op  = const id,
+  arg = const id,
+  var = const id,
+  inp = id,
+  ap  = \f a -> [| $f $a |]
 }
 
 -- | Derive the instance with the given `Derivator` and the given instance head.
@@ -85,26 +83,27 @@ deriveInstance :: Derivator -> Q Type -> Q [Dec]
 deriveInstance deriv qtyp = do
   typ <- qtyp
   case typ of
-    ForallT _ ctx (AppT (ConT className) typeName) ->
-      deriveInstance' deriv ctx className typeName
-    AppT (ConT className) typeName ->
-      deriveInstance' deriv [] className typeName
+    ForallT _ ctx (AppT ty typeName) ->
+      deriveInstance' deriv ctx (className ty) (AppT ty typeName)
+    AppT ty typeName ->
+      deriveInstance' deriv [] (className ty) (AppT ty typeName)
     _ -> fail $ "No support for type: " ++ show typ
 
 deriveInstance' :: Derivator -> Cxt -> Name -> Type -> Q [Dec]
-deriveInstance' deriv ctx className typeName = do
-  ClassI (ClassD _ _ [KindedTV tvn _] _ decs) _ <- reify className
+deriveInstance' deriv ctx clsName typ = do
+  ClassI (ClassD _ _ tvs _ decs) _ <- reify clsName
+  let KindedTV tvn _ = last tvs
   impl <- for decs $ \case
     SigD nm tp -> do
       dec <- reify nm
       case dec of
         ClassOpI{} -> do
           (argNames, body) <- buildOperation deriv tvn tp (op deriv nm (varExp nm))
-          let args = map (\argName -> if contains argName body then VarP argName else WildP) argNames
+          args <- traverse (\argName -> if contains argName body then (if nameBase argName == "var" then inp deriv else id) . pure . VarP $ argName else pure WildP) argNames
           pure $ Just $ FunD nm [Clause args (NormalB body) []]
         _ -> fail $ "No support for declaration: " ++ show dec
     _ -> pure Nothing
-  pure [InstanceD Nothing ctx (AppT (ConT className) typeName) $ catMaybes impl]
+  pure [InstanceD Nothing ctx typ $ catMaybes impl]
 
 buildOperation :: Derivator -> Name -> Type -> Q Exp -> Q ([Name], Exp)
 buildOperation d nm (AppT (AppT ArrowT h) t) e | hasVar nm h = do
@@ -116,8 +115,9 @@ buildOperation d nm (AppT (AppT ArrowT h) t) e = do
   (args, rhs) <- buildOperation d nm t (ap d e (arg d h (varExp varNm)))
   pure (varNm : args, rhs)
 buildOperation d nm (ForallT _ _ t) e = buildOperation d nm t e
+buildOperation d nm (AppT _ t) e | isVar nm t = ([],) <$> eff d e
 buildOperation d nm t e | isVar nm t = ([],) <$> res d e
-                        | otherwise = ([],) <$> e
+                        | otherwise = ([],) <$> cst d e
 
 buildArgument :: Name -> Type -> Q Exp -> Q Exp -> Q Exp
 buildArgument nm (AppT h _) _ var | isVar nm h = var
@@ -134,9 +134,10 @@ hasVar nm (VarT nm') = nm == nm'
 hasVar nm (AppT f a) = isVar nm f || hasVar nm a
 hasVar _ _ = False
 
-tvName :: TyVarBndr -> Name
-tvName (PlainTV nm) = nm
-tvName (KindedTV nm _) = nm
+className :: Type -> Name
+className (ConT nm) = nm
+className (AppT h _) = className h
+className t = error $ "No support for class: " ++ show t
 
 contains :: Data d => Name -> d -> Bool
 contains nm = gmapQl (||) False (\d -> maybe (contains nm d) (== nm) $ cast d)