+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+v0.2   2020-06-29
+-----------------
+  - Support deriving co- and dialgebraic classes with `cst` and `eff`
+  - Support doing pattern matching for input variables with `inp`
+
 v0.1.1 2020-06-24
 -----------------
   - Add `monoidDerivBy`